chore: convert weather simulator to TypeScript
