+++ backend/package.json
@@ -2,12 +2,12 @@
   "name": "climate-event-ohlc-aggregator",
   "version": "1.0.0",
   "description": "Real-time service that connects to a climate WebSocket stream, aggregates data into OHLC format, and serves it via a local API.",
-  "main": "src/index.js",
+  "main": "dist/index.js",
   "scripts": {
     "start": "node dist/index.js",
     "build": "tsc",
-    "dev": "nodemon",
-    "simulate": "node scripts/weather-simulator.js",
+    "dev": "nodemon --watch src --ext ts --exec ts-node src/index.ts",
+    "simulate": "ts-node src/scripts/weather-simulator.ts",
     "format": "prettier --write .",
     "lint": "eslint src scripts utils"
   },
@@ -35,12 +35,14 @@
     "@types/express": "^5.0.3",
     "@types/morgan": "^1.9.10",
     "@types/node": "^24.0.14",
+    "@types/node-fetch": "^2.6.12",
     "@types/ws": "^8.18.1",
     "eslint": "^9.30.1",
     "eslint-config-prettier": "^10.1.5",
     "eslint-plugin-prettier": "^5.5.1",
     "nodemon": "^3.1.10",
     "prettier": "^3.6.2",
+    "ts-node": "^10.9.2",
     "typescript": "^5.8.3"
   },
   "engines": {
